requirements.txt
asgiref==3.5.2
autopep8==1.6.0
black==22.6.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.50.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==21.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
isort==5.10.1
Markdown==3.3.7
mypy-extensions==0.4.3
mysqlclient==2.1.0
oauthlib==3.2.0
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyotp==2.6.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
virtualenv==20.14.1
whitenoise==6.2.0
zipp==3.8.0
ckeditors
psycopg
dj-database-url